agent/requirements.txt
openai
pandas
numpy
sentence-transformers
faiss-cpu
pillow
pytesseract
python-dotenv